selector/requirements.txt
fastapi==0.115.13
huggingface_hub==0.30.2
numpy==1.26.4
openai==1.88.0
peft==0.15.2
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.0
scikit_learn==1.2.2
torch==2.5.1
transformers==4.52.4
uvicorn==0.34.3
sentence-transformers==3.3.0